Remove `-c` CLI option
