+++ tests/test_cli.py
@@ -239,7 +239,6 @@ class TestCli(unittest.TestCase):
     @parameterized.expand([
         ("--config-file", ),
         ("-f", ),
-        ("-c", ),
     ])
     def test_ping(self, config_file_option):
         config_file = 'dwave.conf'
@@ -285,7 +284,6 @@ class TestCli(unittest.TestCase):
     @parameterized.expand([
         ("--config-file", ),
         ("-f", ),
-        ("-c", ),
     ])
     def test_sample(self, config_file_option):
         config_file = 'dwave.conf'
@@ -332,7 +330,6 @@ class TestCli(unittest.TestCase):
     @parameterized.expand([
         ("--config-file", ),
         ("-f", ),
-        ("-c", ),
     ])
     def test_solvers(self, config_file_option):
         config_file = 'dwave.conf'
@@ -373,7 +370,6 @@ class TestCli(unittest.TestCase):
     @parameterized.expand([
         ("--config-file", ),
         ("-f", ),
-        ("-c", ),
     ])
     def test_upload(self, config_file_option):
         config_file = 'dwave.conf'
